Initialize an NMDC client

Initialize an NMDC client with the specified parameters.

NmdcClient$new(
  base_url = "https://api.microbiomedata.org",
  max_page_size = 1000,
  sleep_seconds = 0.5,
  timeout = 30,
  verbose = FALSE,
  links = NULL)

Arguments

base_url

The base URL of the NMDC API.

max_page_size

The maximum number of objects to retrieve per request.

sleep_seconds

The number of seconds to wait between API requests when paginating.

timeout

The number of seconds to wait for an API request to complete before throwing an exception.

verbose

Whether to print URLs of API requests.

links

A data frame of links between objects in the NMDC schema. Defaults to NULL which fetches and caches the links on the first call to `related_ids` or `merge_related`.

Value

An instance of the `NmdcClient` class.