Find related object identifiers of the specified class for the specified IDs

This function finds related objects of the specified class for the specified IDs, and returns a data frame mapping each source ID to a related target ID.

client <- NmdcClient$new()
client$related_ids(ids, cls)

Arguments

ids

A list of object identifiers.

cls

The type of objects to find related objects for.

Value

A data frame mapping each source ID to a related target ID.

Examples

related <- client$related_ids(c('nmdc:bsm-13-7qxjvr77'), nmdcr::WorkflowExecution)
#> Error: object 'client' not found